nvallot/requirements.txt
protobuf==3.20.3
azure-core==1.24.0
azureml-mlflow==1.50.0
azureml-core==1.50.0
azure-identity==1.7.0
azureml-defaults==1.50.0
azure-storage-blob==12.10.0
azure-ai-ml==1.8.0
opencensus-ext-azure==1.1.9
azureml==0.2.7
mlflow==2.5.0
azure-mgmt-core==1.3.2
azureml-inference-server-http==0.8.0
azure-mgmt-core==1.3.2
azure-core==1.24.0
msrest>=0.7.1
azureml-telemetry==1.50.0
azureml-train==1.50.0
azureml-train-core==1.50.0
keras
tensorflow
azure-mgmt-machinelearningservices==2.0.0b1
